+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Changes in 0.8.X
 - Added support for multiple slices and made slicing chainable. (#170) (#190) (#191)
 - Fixed GridFSProxy __getattr__ behaviour (#196)
 - Fix Django timezone support (#151)
+- Simplified Q objects, removed QueryTreeTransformerVisitor (#98) (#171)
 
 Changes in 0.7.9
 ================